Update windmodels
Review of the windmodels revealed problems in defining variables that rely
on latitude in the Kepert wind field. This change resolves the conflicts
that resulted in anomalous wind speeds for southern hemisphere TCs. We
also implement a tempering of the magnitude of the asymmetry that is
intensity dependent, and also decays with increasing distance from the TC
core.