Change datetime variable to double
Identified the datetime variable was not correctly storing float values
as a standard float value, which resulted in mangled datetime values on
extraction from the netCDF file.

In some contexts, the datetime field is not the second record in the
recarray, so the change of dtype was mangling the wrong field. Now
dynamically identifies the correct field based on the name.