Correctly working singleton for configuration
Update the way the 'singleton' decorator works. Now aims to result in a
single state, not a single instance. The previous design pattern caused
problems in testing (we maintained a single instance of `config`
throughout the entire test suite, which led to test failure when running
test_config - the `config` instance was instantiated in KDEOrigin.