Update old-style cnfGetIniValue
If default value was missing, then it immediately did a geteval(). This
caused SyntaxErrors when the value was a string containing operators (e.g.
'/' as found in paths). Now falls back to a get() only if the geteval()
raises a SyntaxError. Also handles the case of a trying to read from a
section that is absent from the configuration file.