Add working singleton class
THis implements a more rigorous singleton mixin for handling configuration
settings. There were applications where the previous singleton approach
would return new instances of ConfigParser (notebooks). This version
provides a transparent method for creating instances of ConfigParser, so
no changes are actually required to the rest of the code base.