New radius to maximum wind model
Radius to maximum wind is a function of pressure deficit and latitude,
after Powell et al (2005). Deault values for the coefficients in the
model are based on southern hemisphere data from 2002 - 2014
(JTWC). Includes a function to fit Rmax data (where available) to the
model (e.g. for application in other domains).