Refactor database generation
Some basic profiling shows that loading the netcdf files is time-consuming
in generating the database. Formerly, :meth:`database.generateEventTable`
and :class:`database.processEvents` both opened each netCDF file. This was
updated to parse a :class:`netCDF4.Dataset` instance to the separate
functions that extract information for the two methods. This means that
:meth:`database.generateEventTable` has been subsumed into
:meth:`database.processEvents`.