Port Hedland example config file
This is the most basic example supplied with TCRM. It's useful to test the
model has been installed correctly and runs without error.

The simulation is a 20 degree by 20 degree domain centred on Port Hedland,
Western Australia. 1000 years of TCs are simulated, and hazard is
calculated and plotted. It uses IBTrACS v03r06 as input.