Boundary layer model hemispheric fix
The Kepert linear boundary layer model required modifications to account
for the different sign of tangential wind speed and vorticity. This commit
makes the output of the KepertWindField identical in northern and southern
hemispheres (aside from a flip in the N-S direction). Note this does not
resolve the issue associated with highly-peaked radial profiles
(especially using the Holland profile).