Correct attributes for time interval
Set the correct interval for the assumed time interval in the netCDF
files, using the cell_methods attribute to describe the time averaging
period for the wind speed variable. This attribute is a bit confusing,
since it says "maximum (interval: 0.2 seconds)". This is interpreted as
the maximum wind speed, where the wind speed is measured over a 0.2 second
interval.