Fix for sqlite3.ProgrammingError
When running the hazard plotting routines immediately after creating the
database, a ProgrammingError was raised. This was caused by the plotting
routines trying to query the existing database instance, which is closed
after creating and populating the database. This commit turns off the
explicit call to `sqlite3.conn.close` and leaves closing the database to
the exit handler. Closes #50.