New notebook to plot landfall intensity distribution
This notebook plots the rate of landfalling TC events, based on TCRM track files, and a set of gates located around a coastline.
The landfalling segments of each track are categorised based on central pressure at the time of landfall, and the mean number of TCs in each category is plotted. We also plot the mean landfall intensity and the minimum landfall intensity of all the events sampled.
This should be more readily transferable to other basins, so long as the gate orientation is chosen appropriately, such that "offshore" areas are to the topological left of each gate segment.