Initial translation to mpi4py
This is the first step in translating parallelisation to mpi4py. It requires
a dummy mpi4py instance to pass through any functions if mpi4py
is not available.

This eliminates the dependency on `pypar`.

TODO: The changes to the `Evaluate` module have not yet been implemented
but will follow the same pattern as in `wind`, `TrackGenerator`,
`hazard` and `database` modules.