Reduce occurrence of logging messages from external modules
When logging level is set to DEBUG, a heap of matplotlib, shapely and other external modules generate logging messages, making it difficult to identify issues with the TCRM code base and issues with other code.

The changes to the `wind.writer` class also fix an issue with temporary files on Windows OS where the OS does not release the temporary file during test cleanup operations.